restaurant-recommendation/phase-1-data-pipeline/requirements.txt
datasets>=2.14.0
pandas>=2.0.0
sqlalchemy>=2.0.0
pytest>=7.4.0
python-dotenv>=1.0.0
huggingface_hub>=0.16.0
